// tools/conformance/src/modules/react-component-generator/react-component-generator.types.ts
// 🏷️ Types

import type { Tree } from "@nx/devkit";

/**
 * Auto-generated documentation placeholder.
 */
export interface ReactComponentGeneratorArguments {
  options: ReactComponentGeneratorOptions;
  tree: Tree;
}

/**
 * Auto-generated documentation placeholder.
 */
export interface ReactComponentGeneratorOptions {
  name: string;
  project?: string;
}